Updated Assignment-9 with naming conventions
