+++ Week-9/todolist/src/Components/TodoList.js
@@ -1,39 +1,23 @@
-import React,{useState} from 'react'
-import Todoform from './Todoform'
-import Todo from './Todo';
+import React from "react";
+import { RiCloseCircleLine } from "react-icons/ri";
 
-function TodoList() {
-    const [todos,setTodo] = useState([]); // used state here to get todo list state
-    const addTodo = todo =>{ // this function add todo into the updated state
-        if(!todo.text){
-            return;
-        }
-        const newTodo = [todo,...todos];
-        setTodo(newTodo);
-    }
-
-    const removeTodo = id =>{ // this function removes a particular todo from list
-        const removeArr = [...todos].filter(todo=>todo.id!==id);
-        setTodo(removeArr);
-    }
-
-    const completeTodo = id =>{ // this function change the pending state to complete by just toggle
-        let updatedTodos = todos.map(todo=>{
-            if(todo.id === id){
-                todo.isComplete = !todo.isComplete;
-            }
-            return todo;
-        });
-        setTodo(updatedTodos);
-    }
-
-  return (
-    <div>
-        <h1>To Do List</h1>
-        <Todoform onSubmit={addTodo}/>
-        <Todo todos={todos} completeTodo = {completeTodo} removeTodo = {removeTodo} />
+function TodoList({ todos, checkedTodo, removeTodo }) {
+  
+  return todos.map((todo, index) => (
+    <div className="todo_row" key={index}>
+      <div key={todo.id}>{todo.text}</div>
+      <div className="icons">
+      <div className={todo.flag ? "status_text_complete" : "status_text_pending"} /* this tells wheather the state is pending or completed */ >
+        {todo.flag ? "Completed" : "Pending"} 
+        <input className="checkBox" type="checkbox" onClick={()=>checkedTodo(todo.id)} defaultChecked={todo.flag? true:false}/* this toggles the state between pending or completed */ />
+      </div>
+        <RiCloseCircleLine
+          onClick={() => removeTodo(todo.id)} //removes a particular todo from list
+          className="delete_icon"
+        />
+      </div>
     </div>
-  )
+  ));
 }
 
 export default TodoList;
